MSVC: increased warning level from 2 to 3.
git-svn-id: http://svn.tilp.info/repos/tiemu@331 4f589446-7ac1-0310-82c7-a055b08196ef

+++ tiemu/trunk/src/core/ti_hw/debugger.c
@@ -230,7 +230,7 @@ int enter_debugger(void)
 	  while(len > 0) 
 	    {
 	      printf("\n0x%08x:", mempos);
-	      for(i=0; i<(len > 16 ? 16 : len); i++)
+	      for(i=0; i<(len > 16 ? 16 : (int)len); i++)
 		printf(" %02x", (*ptr++)&0xff);
 	      mempos+=16;
 	      len-=16;
